backend/requirements.txt
fastapi
uvicorn
pandas
requests
python-dotenv
web3
pycryptodome
